Add more checks on restricted guests

+++ src/main/java/org/linagora/linshare/core/business/service/impl/GuestBusinessServiceImpl.java
@@ -43,6 +43,7 @@ import org.linagora.linshare.core.domain.entities.Account;
 import org.linagora.linshare.core.domain.entities.AllowedContact;
 import org.linagora.linshare.core.domain.entities.Guest;
 import org.linagora.linshare.core.domain.entities.User;
+import org.linagora.linshare.core.exception.BusinessErrorCode;
 import org.linagora.linshare.core.exception.BusinessException;
 import org.linagora.linshare.core.repository.AllowedContactRepository;
 import org.linagora.linshare.core.repository.GuestRepository;
@@ -135,7 +136,9 @@ public class GuestBusinessServiceImpl implements GuestBusinessService {
 		guest.setPassword(hashedPassword);
 		Guest create = guestRepository.create(guest);
 		if (create.isRestricted()) {
-			if (allowedContacts != null) {
+			if (allowedContacts == null || allowedContacts.isEmpty()) {
+				throw new BusinessException(BusinessErrorCode.GUEST_INVALID_INPUT, "You can not create a restricted guest without a list of contacts.");
+			} else {
 				for (User contact : allowedContacts) {
 					allowedContactRepository.create(new AllowedContact(create,
 							contact));
@@ -166,6 +169,9 @@ public class GuestBusinessServiceImpl implements GuestBusinessService {
 		Guest update = guestRepository.update(entity);
 		if (wasRestricted == guest.isRestricted()) {
 			if (allowedContacts != null) {
+				if (allowedContacts.isEmpty()) {
+					throw new BusinessException(BusinessErrorCode.GUEST_INVALID_INPUT, "You can not update a restricted guest without a list of contacts.");
+				}
 				// update
 				allowedContactRepository.purge(update);
 				for (User contact : allowedContacts) {
@@ -174,13 +180,20 @@ public class GuestBusinessServiceImpl implements GuestBusinessService {
 				}
 			}
 		} else if (wasRestricted) {
-			// not restricted anymore. purge
+			// it is not restricted anymore. purge
 			allowedContactRepository.purge(update);
 		} else {
-			// Add
-			for (User contact : allowedContacts) {
-				allowedContactRepository.create(new AllowedContact(update,
-						contact));
+			// it was not restricted,
+			if (guest.isRestricted()) {
+				// but it is now
+				if (allowedContacts == null || allowedContacts.isEmpty()) {
+					throw new BusinessException(BusinessErrorCode.GUEST_INVALID_INPUT, "You can not update a restricted guest without a list of contacts.");
+				} else {
+					for (User contact : allowedContacts) {
+						allowedContactRepository.create(new AllowedContact(update,
+								contact));
+					}
+				}
 			}
 		}
 		return update;
